+++ clean_architecture/lib/logic/domain/entities/employee.dart
@@ -1,7 +1,7 @@
 abstract class Employee {
   final int id;
   final String employeeName;
-  final double employeeSalary;
+  final int employeeSalary;
   final int employeeAge;
 
   /// There is another field from the response, namely [profile_image] but I have omitted it, as it is always empty
